Deduplicate and move lastDay tests to utils package
