+++ utils/utils_test.go
@@ -1,0 +1,37 @@
+/*
+Copyright 2017 - The GoMiler Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package utils
+
+import (
+	"testing"
+	"time"
+)
+
+func TestLastDayMonth(t *testing.T) {
+	date := time.Now().Local()
+	lastDay := LastDayMonth(date.Year(), int(date.Month()), time.UTC)
+	expectedDay := time.Date(date.Year(), time.Month(date.Month())+1, 0, 0, 0, 0, 0, time.UTC)
+	if lastDay != expectedDay {
+		t.Errorf("Expected %v, got %v", expectedDay, lastDay)
+	}
+}
+
+func TestLastDayWeek(t *testing.T) {
+	date := time.Now().Local()
+	lastDay := LastDayWeek(date)
+	if lastDay.Weekday() != time.Sunday {
+		t.Errorf("Expected %s, got %s", time.Sunday, lastDay.Weekday())
+	}
+}
